feat(121): restructure roadmap components
