+++ apps/xi.land/components/roadmap/index.ts
@@ -1,0 +1,2 @@
+export { overviewCards, roadmapData } from './dataForRoadmap';
+export { RoadmapTimeline } from './RoadmapTimeline';
